backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
groq>=0.4.0
pinecone>=5.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tavily-python>=0.5.0
pydantic>=2.5.0
